backend/requirements.txt
openai
langchain
langchain-openai
langchain-community
faiss-cpu
llama-cloud-services
llama-index-core
llama-index-readers-file
python-dotenv
fastapi
uvicorn
python-multipart
pydantic
langchain-text-splitters
streamlit
pandas
PyMuPDF
python-magic
pytest